[REFACTOR] Load tutorial contents lazily
